requirements.txt
pytest
pytorch
numpy
networkx
matplotlib
